Get back to a set of functional tests
