+++ apps/test/unit/lib/levelbuilder/script-editor/LessonGroupCardTest.jsx
@@ -5,13 +5,11 @@ import {UnconnectedLessonGroupCard as LessonGroupCard} from '@cdo/apps/lib/level
 
 const defaultProps = {
   lessonGroupsCount: 1,
-  setLessonMetrics: () => {},
-  setTargetLesson: () => {},
-  targetLessonPos: 1,
-  lessonMetrics: {},
   addLesson: () => {},
   moveGroup: () => {},
   removeGroup: () => {},
+  moveLesson: () => {},
+  removeLesson: () => {},
   convertGroupToUserFacing: () => {},
   lessonGroup: {
     key: 'lg-key',
@@ -22,41 +20,14 @@ const defaultProps = {
       {
         id: 100,
         name: 'A',
-        position: 1,
-        levels: [
-          {
-            ids: [1],
-            position: 1,
-            activeId: 1
-          },
-          {
-            ids: [4],
-            position: 2,
-            activeId: 4
-          },
-          {
-            ids: [5],
-            position: 3,
-            activeId: 5
-          },
-          {
-            ids: [6],
-            position: 4,
-            activeId: 6
-          }
-        ]
+        key: 'lesson-1',
+        position: 1
       },
       {
         name: 'B',
+        key: 'lesson-2',
         id: 101,
-        position: 2,
-        levels: [
-          {
-            ids: [2, 3],
-            position: 1,
-            activeId: 3
-          }
-        ]
+        position: 2
       }
     ]
   }
@@ -66,7 +37,7 @@ describe('LessonGroupCard', () => {
   it('displays LessonGroupCard correctly', () => {
     let wrapper = shallow(<LessonGroupCard {...defaultProps} />);
     expect(wrapper.find('OrderControls')).to.have.lengthOf(1);
-    expect(wrapper.find('Connect(UnconnectedLessonCard)')).to.have.lengthOf(2);
+    expect(wrapper.find('LessonToken')).to.have.lengthOf(2);
     expect(wrapper.find('button')).to.have.lengthOf(1);
   });
 });
